+++ creme/metrics/cross_entropy.py
@@ -59,6 +59,9 @@ class CrossEntropy(stats.Mean, BaseCrossEntropy):
 class RollingCrossEntropy(stats.RollingMean, BaseCrossEntropy):
     """Rolling cross-entropy.
 
+    Parameters:
+        window_size (int): Size of the window of recent values to consider.
+
     Example:
 
         ::